+++ deepface/modules/preprocessing.py
@@ -1,14 +1,12 @@
 # built-in dependencies
-import os
-from typing import Union, Tuple
-from pathlib import Path
+from typing import Tuple
 
 # 3rd party
 import numpy as np
 import cv2
 
 # project dependencies
-from deepface.commons import package_utils, file_utils
+from deepface.commons import package_utils
 
 
 tf_major_version = package_utils.get_tf_major_version()
@@ -16,49 +14,6 @@ if tf_major_version == 1:
     from keras.preprocessing import image
 elif tf_major_version == 2:
     from tensorflow.keras.preprocessing import image
-
-
-def load_image(img: Union[str, np.ndarray]) -> Tuple[np.ndarray, str]:
-    """
-    Load image from path, url, base64 or numpy array.
-    Args:
-        img: a path, url, base64 or numpy array.
-    Returns:
-        image (numpy array): the loaded image in BGR format
-        image name (str): image name itself
-    """
-
-    # The image is already a numpy array
-    if isinstance(img, np.ndarray):
-        return img, "numpy array"
-
-    if isinstance(img, Path):
-        img = str(img)
-
-    if not isinstance(img, str):
-        raise ValueError(f"img must be numpy array or str but it is {type(img)}")
-
-    # The image is a base64 string
-    if img.startswith("data:image/"):
-        return file_utils.load_base64(img), "base64 encoded string"
-
-    # The image is a url
-    if img.lower().startswith("http://") or img.lower().startswith("https://"):
-        return file_utils.load_image_from_web(url=img), img
-
-    # The image is a path
-    if os.path.isfile(img) is not True:
-        raise ValueError(f"Confirm that {img} exists")
-
-    # image must be a file on the system then
-
-    # image name must have english characters
-    if img.isascii() is False:
-        raise ValueError(f"Input image must not have non-english characters - {img}")
-
-    img_obj_bgr = cv2.imread(img)
-    # img_obj_rgb = cv2.cvtColor(img_obj_bgr, cv2.COLOR_BGR2RGB)
-    return img_obj_bgr, img
 
 
 def normalize_input(img: np.ndarray, normalization: str = "base") -> np.ndarray:
