+++ tests/test_find.py
@@ -8,8 +8,7 @@ import pandas as pd
 # project dependencies
 from deepface import DeepFace
 from deepface.modules import verification
-from deepface.modules import recognition
-from deepface.commons import file_utils
+from deepface.commons import image_utils
 from deepface.commons import logger as log
 
 logger = log.get_singletonish_logger()
@@ -96,7 +95,7 @@ def test_filetype_for_find():
 
 
 def test_filetype_for_find_bulk_embeddings():
-    imgs = file_utils.list_images("dataset")
+    imgs = image_utils.list_images("dataset")
 
     assert len(imgs) > 0
